feat(filters): cleanup
