+++ packages/falcon-ecommerce-uikit/src/Category/SortOrderDropdown.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Text, Box, Dropdown, DropdownLabel, DropdownMenu, DropdownMenuItem, themed } from '@deity/falcon-ui';
 import { T } from '@deity/falcon-i18n';
-import { SortOrder } from './SortOrdersQuery';
+import { SortOrder } from '../SortOrders/SortOrdersQuery';
 
 export const SortOrderDropdownLayout = themed({
   tag: Box,
