+++ packages/falcon-ecommerce-uikit/src/Category/index.ts
@@ -1,6 +1,5 @@
 export * from './CategoryProductsQuery';
 export * from './CategoryLayout';
-export * from './SortOrdersQuery';
 export * from './SortOrderDropdown';
 export * from './ShowingOutOf';
 export * from './ShowMore';
