+++ packages/falcon-ecommerce-uikit/src/Filters/FilterItem.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Icon, ListItem, themed } from '@deity/falcon-ui';
-import { AggregationBucket } from '../Search/types';
+import { AggregationBucket } from '../Category/CategoryProductsQuery';
 
 export const FilterItemLayout = themed({
   tag: ListItem,
