+++ packages/falcon-ecommerce-uikit/src/Search/types.ts
@@ -1,5 +1,18 @@
 import { PaginationInput } from './../types';
-import { SortOrderInput } from './../Category/SortOrdersQuery';
+import { SortOrderInput } from '../SortOrders/SortOrdersQuery';
+
+export type SearchState = {
+  term?: string;
+  filters: FilterInput[];
+  sort: SortOrderInput;
+  pagination?: PaginationInput;
+};
+
+export type FilterInput = {
+  field: string;
+  operator: FilterOperator;
+  value: string[];
+};
 
 const FilterOperatorValues = {
   eq: 'eq',
@@ -17,45 +30,3 @@ type FilterOperatorsType = Record<FilterOperator, FilterOperator>;
 export const FilterOperators: FilterOperatorsType = Object.freeze<FilterOperatorsType>({
   ...(FilterOperatorValues as FilterOperatorsType)
 });
-
-export type FilterInput = {
-  field: string;
-  operator: FilterOperator;
-  value: string[];
-};
-
-export type SearchState = {
-  term?: string;
-  filters: FilterInput[];
-  sort: SortOrderInput;
-  pagination?: PaginationInput;
-};
-
-export type Aggregation = {
-  field: string;
-  type: SelectionType;
-  buckets: AggregationBucket[];
-  title: string;
-};
-
-export type SelectionType = 'single' | 'multiple' | 'range';
-
-export type AggregationBucket = {
-  value: string;
-  count: number;
-  title: string;
-};
-
-export type FilterData = {
-  field: string;
-  title: string;
-  type: SelectionType;
-  options: FilterOption[];
-  value: string[];
-};
-
-export type FilterOption = {
-  title: string;
-  value: string;
-  count: number;
-};
