+++ packages/falcon-ecommerce-uikit/src/SortOrders/index.ts
@@ -1,0 +1,2 @@
+export * from './SortOrdersQuery';
+export * from './SortOrdersProvider';
