+++ packages/falcon-ecommerce-uikit/src/index.ts
@@ -30,7 +30,6 @@ export * from './Sidebar';
 export * from './Forms';
 export * from './AccountRecovery';
 export * from './Errors';
-export * from './SortOrdersProvider';
 export * from './Search';
 export * from './OnlineStatus';
 export * from './EnsureTTI';
@@ -38,3 +37,4 @@ export * from './BackendConfig';
 export * from './Filters';
 export * from './Responsive';
 export * from './TwoStepWizard';
+export * from './SortOrders';
